+++ src/main/groovy/life/qbic/samplestatus/reporter/QbicSampleStatusReporter.groovy
@@ -7,6 +7,8 @@ import life.qbic.samplestatus.reporter.api.SampleTrackingService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
 
+import java.time.Instant
+
 /**
  * <b>Reports a sample update to the sample-tracking-service</b>
  * @since 1.0.0
@@ -29,9 +31,10 @@ class QbicSampleStatusReporter implements SampleStatusReporter {
         })
         String sampleCode = sampleUpdate.getSample().getSampleCode()
         String status = sampleUpdate.getUpdatedStatus()
+        Instant updateTimepoint = sampleUpdate.getModificationDate()
         Person responsiblePerson = locationService.getResponsiblePerson().orElseThrow({
             new RuntimeException("No responsible person for the update was determined.")
         })
-        sampleTrackingService.updateSampleLocation(sampleCode, currentLocation, status, responsiblePerson)
+        sampleTrackingService.updateSampleLocation(sampleCode, currentLocation, status, updateTimepoint, responsiblePerson)
     }
 }