+++ src/main/groovy/life/qbic/samplestatus/reporter/services/NcctLocationService.groovy
@@ -1,11 +1,10 @@
 package life.qbic.samplestatus.reporter.services
 
 import life.qbic.samplestatus.reporter.api.Location
-import life.qbic.samplestatus.reporter.api.Person
-import life.qbic.samplestatus.reporter.services.users.UserService
 import life.qbic.samplestatus.reporter.api.LocationService
+import life.qbic.samplestatus.reporter.api.Person
 import life.qbic.samplestatus.reporter.api.SampleTrackingService
-
+import life.qbic.samplestatus.reporter.services.users.UserService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.boot.context.properties.ConfigurationProperties
@@ -38,7 +37,7 @@ class NcctLocationService implements LocationService {
 
     @Override
     Optional<Person> getResponsiblePerson() {
-        Optional<Person> responsiblePerson = userService.getUserDetails(userId)
+        Optional<Person> responsiblePerson = userService.getPerson(userId)
         return responsiblePerson
     }
 }